tp09 - dijkstra fonctionnel

+++ src/Interfaces/GraphInterface.java
@@ -14,6 +14,6 @@ public interface GraphInterface {
 	//Interface retenue
     public ArrayList<VertexInterface> getAllVertices() ; //donne la liste des sommets
     public ArrayList<VertexInterface> getSuccessors(VertexInterface vertex) ; //donne la liste des successeurs de vertex
-    public int getWeight(VertexInterface src,VertexInterface dst) ; //donne la distance entre s1 et s2
+    public Double getWeight(VertexInterface src,VertexInterface dst) ; //donne la distance entre s1 et s2
     
 }
